Send CreateShareLinkRequest in class AdvancedSharedLinkDialog internal.
